Sort and format
