+++ http4s-quickstart/src/main/scala/com/github/kkrull/http4s/HelloWorld.scala
@@ -17,12 +17,14 @@ object HelloWorld {
     * this shows how you can create encoders for your data.
     */
   final case class Greeting(greeting: String) extends AnyVal
+
   object Greeting {
     implicit val greetingEncoder: Encoder[Greeting] = new Encoder[Greeting] {
       final def apply(a: Greeting): Json = Json.obj(
         ("message", Json.fromString(a.greeting)),
       )
     }
+
     implicit def greetingEntityEncoder[F[_]]: EntityEncoder[F, Greeting] =
       jsonEncoderOf[F, Greeting]
   }
