--- http4s-quickstart/src/main/scala/com/github/kkrull/http4s/Http4squickstartRoutes.scala
@@ -1,31 +1,0 @@
-package com.github.kkrull.http4s
-
-import cats.effect.Sync
-import cats.implicits._
-import org.http4s.HttpRoutes
-import org.http4s.dsl.Http4sDsl
-
-object Http4squickstartRoutes {
-
-  def jokeRoutes[F[_]: Sync](J: Jokes[F]): HttpRoutes[F] = {
-    val dsl = new Http4sDsl[F] {}
-    import dsl._
-    HttpRoutes.of[F] { case GET -> Root / "joke" =>
-      for {
-        joke <- J.get
-        resp <- Ok(joke)
-      } yield resp
-    }
-  }
-
-  def helloWorldRoutes[F[_]: Sync](H: HelloWorld[F]): HttpRoutes[F] = {
-    val dsl = new Http4sDsl[F] {}
-    import dsl._
-    HttpRoutes.of[F] { case GET -> Root / "hello" / name =>
-      for {
-        greeting <- H.hello(HelloWorld.Name(name))
-        resp <- Ok(greeting)
-      } yield resp
-    }
-  }
-}
