+++ http4s-quickstart/src/main/scala/com/github/kkrull/http4s/Jokes.scala
@@ -19,19 +19,19 @@ object Jokes {
   def apply[F[_]](implicit ev: Jokes[F]): Jokes[F] = ev
 
   final case class Joke(joke: String) extends AnyVal
+
   object Joke {
     implicit val jokeDecoder: Decoder[Joke] = deriveDecoder[Joke]
-    implicit def jokeEntityDecoder[F[_]: Concurrent]: EntityDecoder[F, Joke] =
-      jsonOf
+    implicit def jokeEntityDecoder[F[_]: Concurrent]: EntityDecoder[F, Joke] =      jsonOf
     implicit val jokeEncoder: Encoder[Joke] = deriveEncoder[Joke]
-    implicit def jokeEntityEncoder[F[_]]: EntityEncoder[F, Joke] =
-      jsonEncoderOf
+    implicit def jokeEntityEncoder[F[_]]: EntityEncoder[F, Joke] =      jsonEncoderOf
   }
 
   final case class JokeError(e: Throwable) extends RuntimeException
 
   def impl[F[_]: Concurrent](C: Client[F]): Jokes[F] = new Jokes[F] {
     val dsl = new Http4sClientDsl[F] {}
+
     import dsl._
     def get: F[Jokes.Joke] =
       C.expect[Joke](GET(uri"https://icanhazdadjoke.com/"))
