+++ http4s-quickstart/src/main/scala/com/github/kkrull/http4s/Main.scala
@@ -3,5 +3,5 @@ package com.github.kkrull.http4s
 import cats.effect.{IO, IOApp}
 
 object Main extends IOApp.Simple {
-  val run = Http4squickstartServer.run[IO]
+  val run = Http4sQuickstartServer.run[IO]
 }
